refactor: split Phase 2 driver .h files into .c + .h pairs
Move function implementations from harness.h, uart.h, spi.h,
can_common.h, and fdcan.h into corresponding .c files.
